Create README.md
he Voting Web App is a comprehensive platform designed for efficient election management and participation. This application facilitates seamless admin registration and login processes and features a robust dashboard for overseeing voters and elections. Leveraging the power of Spring Boot and Thymeleaf, it ensures a smooth experience for both administrators and participants.